+++ special-pages/pages/new-tab/app/activity/ActivityInteractionsContext.js
@@ -1,0 +1,8 @@
+import { createContext } from 'preact';
+
+export const ActivityInteractionsContext = createContext({
+    /**
+     * @type {(evt: MouseEvent) => void} event
+     */
+    didClick(event) {},
+});
