+++ special-pages/pages/new-tab/app/activity/constants.js
@@ -4,5 +4,4 @@
 export const DDG_STATS_OTHER_COMPANY_IDENTIFIER = '__other__';
 export const ACTION_ADD_FAVORITE = 'add-favorite';
 export const ACTION_REMOVE_FAVORITE = 'remove-favorite';
-export const ACTION_BURN = 'burn';
 export const ACTION_REMOVE = 'remove';
